--- SlowGames/Assets/Resources/Prefabs/Audio/SlowGameAudioManager.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 52624b6512a705a48934eef0ad8025ce
-timeCreated: 1477494073
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
