--- SlowGames/Assets/Scripts/Audio/SlowGameAudioManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a87aca736c7d20d49be8fa4b14f3b2e6
-timeCreated: 1477493122
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
